// packages/config/src/schemas/environment.ts
// Generated by ts-to-zod
import { z } from 'zod';

import { countryCodesEnum } from './country-codes.js';

export const environmentSchema = z.object({
  countryCode: countryCodesEnum,
  hostName: z.string(),
  hosts: z.object({
    account: z.string(),
    listen: z.string(),
  }),
  locale: z.string(),
  markPlayedThreshold: z.number(),
  minimumAgeToJoin: z.number(),
  objectDbBucket: z
    .literal('web.listen.staging')
    .or(z.literal('web.listen.prod')),
  presetPopoverDisplayFrequency: z.number(),
  presetPopoverMaxTimesDismissed: z.number(),
  supportedCountries: z.array(z.string()),
  terminalId: z.number(),
  territoryCode: z.string(),
  websiteUrl: z.string(),
  registrationOptions: z
    .object({
      zipRegex: z.string(),
      usePostal: z.boolean().optional(),
      zipKeyboard: z
        .union([z.literal('numeric'), z.literal('alphanumeric')])
        .default('numeric'),
    })
    .optional(),
});

export type Environment = z.infer<typeof environmentSchema>;